refactor(components): [col] switch to script-setup syntax (#7971)
