+++ packages/components/col/index.ts
@@ -1,6 +1,6 @@
 import { withInstall } from '@element-plus/utils'
 
-import Col from './src/col'
+import Col from './src/col.vue'
 
 export const ElCol = withInstall(Col)
 export default ElCol
